Add delete Instances to Deployment API and refactoring openshift framework
