+++ framework-cloud/framework-openshift/src/main/java/org/kie/cloud/openshift/deployment/OpenShiftDeployment.java
@@ -1,0 +1,66 @@
+/*
+ * Copyright 2017 JBoss by Red Hat.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.kie.cloud.openshift.deployment;
+
+import java.util.Arrays;
+import java.util.List;
+
+import org.kie.cloud.api.deployment.Deployment;
+import org.kie.cloud.api.deployment.Instance;
+import org.kie.cloud.openshift.OpenShiftController;
+
+public abstract class OpenShiftDeployment implements Deployment {
+
+    protected OpenShiftController openShiftController;
+    protected String namespace;
+
+    public OpenShiftController getOpenShiftController() {
+        return openShiftController;
+    }
+
+    public void setOpenShiftController(OpenShiftController openShiftController) {
+        this.openShiftController = openShiftController;
+    }
+
+    @Override
+    public String getNamespace() {
+        return namespace;
+    }
+
+    public void setNamespace(String namespace) {
+        this.namespace = namespace;
+    }
+
+    @Override
+    public void deleteInstances(Instance... instance) {
+        deleteInstances(Arrays.asList(instance));
+    }
+
+    @Override
+    public void deleteInstances(List<Instance> instances) {
+        for (Instance instance : instances) {
+            openShiftController.getClient().pods().inNamespace(namespace).withName(instance.getName()).withGracePeriod(0).delete();
+        }
+    }
+
+    public abstract String getServiceName();
+
+    @Override
+    public void scale(int instances) {
+        openShiftController.getProject(namespace).getService(getServiceName()).getDeploymentConfig().scalePods(instances);
+    }
+
+}
